+++ rmk-macro/src/flash.rs
@@ -12,8 +12,8 @@ pub(crate) fn expand_flash_init(keyboard_config: &KeyboardTomlConfig) -> TokenSt
     if keyboard_config.storage.as_ref().map_or(true, |s| !s.enabled) {
         // This config actually does nothing if storage is disabled
         return quote! {
-            let storage_config = ::rmk::config::StorageConfig::default();
-            let flash = ::rmk::DummyFlash::new();
+            // let storage_config = ::rmk::config::StorageConfig::default();
+            // let flash = ::rmk::DummyFlash::new();
         };
     }
     let mut flash_init = get_storage_config(keyboard_config.storage.as_ref().unwrap());