+++ rmk/src/lib.rs
@@ -71,6 +71,7 @@ mod boot;
 pub mod channel;
 pub mod combo;
 pub mod config;
+pub mod controller;
 pub mod debounce;
 pub mod direct_pin;
 pub mod event;
@@ -79,7 +80,7 @@ pub mod hid;
 pub mod hid_state;
 pub mod input_device;
 pub mod keyboard;
-mod keyboard_macro;
+pub mod keyboard_macros;
 pub mod keycode;
 pub mod keymap;
 pub mod layout_macro;
@@ -126,7 +127,7 @@ pub async fn initialize_encoder_keymap_and_storage<
     default_keymap: &'a mut [[[KeyAction; COL]; ROW]; NUM_LAYER],
     default_encoder_map: &'a mut [[EncoderAction; NUM_ENCODER]; NUM_LAYER],
     flash: F,
-    storage_config: config::StorageConfig,
+    storage_config: &config::StorageConfig,
     behavior_config: config::BehaviorConfig,
 ) -> (
     RefCell<KeyMap<'a, ROW, COL, NUM_LAYER, NUM_ENCODER>>,
@@ -156,7 +157,7 @@ pub async fn initialize_keymap_and_storage<
 >(
     default_keymap: &'a mut [[[KeyAction; COL]; ROW]; NUM_LAYER],
     flash: F,
-    storage_config: config::StorageConfig,
+    storage_config: &config::StorageConfig,
     behavior_config: config::BehaviorConfig,
 ) -> (
     RefCell<KeyMap<'a, ROW, COL, NUM_LAYER, 0>>,