+++ rmk/CHANGELOG.md
@@ -16,11 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- BLE support for Pi Pico W 
 - Introduce matrix_map for nicer keyboard matrix configs
 - Battery charging state reader
 - Sleep timeout when advertising
-- Manually pairing between central and peripheral
+- Automatically pair between central and peripheral
 - Make constants in RMK configurable via `keyboard.toml`
+- enables support for keyboard macros (via rust based configuration only for now) (closes issues issues #308, #284, #303, #313, #170)
 
 ### Fixed